scottolson/requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
Django==3.0.3
django-debug-toolbar==2.2
django-markdownx==3.0.1
idna==2.8
Markdown==3.2.1
numpy==1.18.1
Pillow==7.0.0
PyJWT==1.7.1
python-dotenv==0.12.0
pytz==2019.3
requests==2.22.0
six==1.14.0
sqlparse==0.3.0
urllib3==1.25.8
